test with new single smoke_tests.js file
