+++ .github/workflows/on_pull_request.yml
@@ -67,25 +67,7 @@ jobs:
       env:
         TEST_API_AUTH_KEY: ${{ secrets.TEST_API_AUTH_KEY }}
       with:
-        filename: smoke_tests/votes_test.js
-    - name: Run k6 delegations smoke test
-      uses: k6io/action@v0.1
-      env:
-        TEST_API_AUTH_KEY: ${{ secrets.TEST_API_AUTH_KEY }}
-      with:
-        filename: smoke_tests/delegations_test.js
-    - name: Run k6 participants smoke test
-      uses: k6io/action@v0.1
-      env:
-        TEST_API_AUTH_KEY: ${{ secrets.TEST_API_AUTH_KEY }}
-      with:
-        filename: smoke_tests/participants_test.js
-    - name: Run k6 voting results smoke test
-      uses: k6io/action@v0.1
-      env:
-        TEST_API_AUTH_KEY: ${{ secrets.TEST_API_AUTH_KEY }}
-      with:
-        filename: smoke_tests/voting_results_test.js
+        filename: smoke_tests/smoke_tests.js
     - name: Login to gcloud
       uses: GoogleCloudPlatform/github-actions/setup-gcloud@master
       with:
